Broke @show into @dump (showRaw) and @show (show).
In some cases it's nice to have slightly more readable output, and
in others it's nice to know exactly which AST nodes are which. Now
we can support both.

Also, we have a nice heuristic for eliminating extra AST junk. When
the user does something like "@show val x = 9" we normally get an
AST like so:

  { val x = 9; () }

org.spirebot.Util.unpack allows us to strip the part we care of out
of the cruft.

Finally, we update to Spire 0.4.0-M1.